+++ packages/widgetbook/CHANGELOG.md
@@ -10,6 +10,7 @@
 - **FEAT**: Support query groups nullification. ([#1664](https://github.com/widgetbook/widgetbook/pull/1664))
 - **FEAT**: Add `SingleFieldOnly` mixin that can be applied to single-field `FieldsComposable` (i.e. `Addon` or `Arg`) to provide default implementations for to/from `QueryGroup`. ([#1670](https://github.com/widgetbook/widgetbook/pull/1670))
 - **FEAT**: Add `NoFields` mixin that can be applied to `FieldsComposable` (i.e. `Addon` or `Arg`) that do not need `Field`s. ([#1672](https://github.com/widgetbook/widgetbook/pull/1672))
+- **FEAT**: Add `path` getter to `Story`. ([#1692](https://github.com/widgetbook/widgetbook/pull/1692))
 - **REFACTOR**: Use `QueryGroup` instead of query parameters and `Map<String, String>`. ([#1653](https://github.com/widgetbook/widgetbook/pull/1653))
 
 ## 4.0.0-alpha.3